smth

+++ app/Http/Controllers/AdminController.php
@@ -12,6 +12,7 @@ use Illuminate\Support\Facades\Mail;
 use App\Mail\PunishmentBan;
 use App\Mail\PunishmentCanceled;
 use App\Mail\PunishmentRecieved;
+use App\Models\Comment;
 
 class AdminController extends Controller
 {
@@ -49,12 +50,60 @@ class AdminController extends Controller
 
         $product = Product::find($productId);
         $punishment = Punishment::where('object_id', $productId)->where('is_product', true)->first();
-        $punishment->delete();
+        if ($punishment) {
+            $punishment->delete();
+        }
 
         $product->is_deleted = false;
         $product->save();
 
         Mail::to(User::find($product->seller_user_id))->send(new PunishmentCanceled($punishment));
+
+        return redirect()->route('products');
+    }
+
+    public function deactivateComment(Request $request, $reviewId)
+    {
+
+        if (Auth::user()->type !== "Admin") {
+            abort(404);
+        }
+
+        $review = Comment::find($reviewId);
+        $punishment = new Punishment();
+
+        $punishment->user_id = $review->user_id;
+        $punishment->object_id = $reviewId;
+        $punishment->reason = $request->reason;
+        $punishment->is_product = false;
+
+        $punishment->save();
+
+        $review->is_deleted = true;
+        $review->save();
+
+        Mail::to(User::find($review->user_id))->send(new PunishmentRecieved($punishment));
+
+        return redirect()->route('products');
+    }
+
+    public function activateComment($reviewId)
+    {
+
+        if (Auth::user()->type !== "Admin") {
+            abort(404);
+        }
+
+        $review = Comment::find($reviewId);
+        $punishment = Punishment::where('object_id', $reviewId)->where('is_product', false)->first();
+        if ($punishment) {
+            $punishment->delete();
+        }
+
+        $review->is_deleted = false;
+        $review->save();
+
+        Mail::to(User::find($review->user_id))->send(new PunishmentCanceled($punishment));
 
         return redirect()->route('products');
     }
